리뷰 컴포넌트 추가 및 ui수정
