--- src/Detail/Components/Review.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Review = () => {
-  return <div></div>;
-};
-
-export default Review;
