+++ src/pages/Detail/Components/Star.scss
