+++ src/pages/Detail/Detail.scss
@@ -110,37 +110,6 @@
   width: 100%;
 }
 
-.reviewPage {
-  height: 40px;
-  width: 100px;
-  border-radius: 6px;
-  border: none;
-  margin: 15px 0px;
-  color: white;
-  font-size: 18px;
-  font-weight: 900;
-  text-align: center;
-  background-color: rgb(0, 150, 243);
-}
-
-.reviewHeader {
-  display: flex;
-  justify-content: space-between;
-  background-color: #f4f4f4;
-  padding: 10px;
-  font-size: 20px;
-  font-family: Arial, Helvetica, sans-serif;
-  font-weight: bold;
-}
-.content {
-  display: flex;
-  padding: 20px 10px;
-  line-height: 30px;
-  font-weight: 600;
-  font-size: 20px;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
 .starIcon {
   display: flex;
   gap: 10px;
